requirements.txt
# ============================================
# TP-LINK Camera Monitor - Python依赖
# ============================================

# 核心依赖
apscheduler==3.10.4
flask==3.0.0
requests==2.31.0
Pillow==10.1.0
psutil==5.9.6

# 可选依赖（根据需要安装）
# opencv-python==4.8.1.78  # 如需视频分析功能
# numpy==1.26.2            # OpenCV依赖
